Use CWM::Dialog
